RavenDB-14487 transaction merger should return long instead of int

+++ src/Raven.Server/Documents/Handlers/Admin/AdminRevisionsHandler.cs
@@ -44,7 +44,7 @@ namespace Raven.Server.Documents.Handlers.Admin
                 _database = database;
             }
 
-            protected override int ExecuteCmd(DocumentsOperationContext context)
+            protected override long ExecuteCmd(DocumentsOperationContext context)
             {
                 foreach (var id in _ids)
                 {
